small refactor of files and interface names
